project/backend/requirements.txt
playwright
playwright-stealth
google-genai
pillow
requests
python-dotenv
psycopg[pool,binary]
pgvector==0.2.5
fastapi
beautifulsoup4
httpx
uvicorn[standard]
supabase
python-multipart
torch
transformers
rembg
